// src/features/payment/types/enums-payments.ts
export enum PaymentStatus {
  PENDING = "PENDING",
  APPROVED = "APPROVED",
  REJECTED = "REJECTED",
  COMPLETED = "COMPLETED",
}

export enum PaymentMethod {
  POINTS = "POINTS",
  CASH = "CASH",
  BANK_TRANSFER = "BANK_TRANSFER",
  CREDIT_CARD = "CREDIT_CARD",
  VOUCHER = "VOUCHER",
  PAYMENT_GATEWAY = "PAYMENT_GATEWAY",
}